refactor(desktop): finish App decomposition — render_dialogs + signal bundle (<1000)
Bundle the App component's ~29 Signal handles into a Copy `AppSignals`
(signals.rs) and move the modal/overlay dialog rendering into
app/dialogs.rs::render_dialogs, which the main rsx! now calls. The
`do_reconnect` reconnect helper becomes a free fn taking the bundle (shared
by the workspace toolbar and the connection dialog); `on_secrets_command`
moves into the dialog renderer.

app/mod.rs is now 982 lines. With this, every source file in the workspace is
under 1000 lines. No behavior change.
